# openapi.json
{"openapi":"3.1.0","info":{"title":"Supplier Product Harmonisation API","description":"Phase 8 Official Catalog & Metrics API Layer exposed for Staging/Production environments.","version":"1.0.0"},"paths":{"/products":{"get":{"summary":"Get all harmonized products with advanced filters","operationId":"get_products_products_get","parameters":[{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter exactly by product category","title":"Category"},"description":"Filter exactly by product category"},{"name":"supplier","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter products supplied by a specific supplier name","title":"Supplier"},"description":"Filter products supplied by a specific supplier name"},{"name":"in_stock","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"True: returns only available variants, False: returns out-of-stock","title":"In Stock"},"description":"True: returns only available variants, False: returns out-of-stock"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Case-insensitive keyword search in product name or description","title":"Search"},"description":"Case-insensitive keyword search in product name or description"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products/{product_id}":{"get":{"summary":"Get a single product by its UUID","operationId":"get_product_by_id_products__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categories":{"get":{"summary":"Get a unique list of all harmonized categories","operationId":"get_categories_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/suppliers":{"get":{"summary":"Get a unique list of all suppliers","operationId":"get_suppliers_suppliers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/metrics":{"get":{"summary":"Get global platform sync metrics","operationId":"get_platform_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/duplicates":{"get":{"summary":"Identify potential product naming duplicates","operationId":"get_potential_duplicates_duplicates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/quality":{"get":{"summary":"Check data harmonisation quality scoring","operationId":"get_data_quality_score_quality_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}